Fixing issue of not returning EINVAL
If a direct I/O write checksum verify error occurred then EINVAL is
suppose to be returned. However, I made the mistake of not propagating
the EINVAL error up to the parents ZIO. In order to fix this, I removed
the ZIO_FLAG_DONT_PROPAGATE from the VDEV child ZIO. This way the EINVAL
error will be set in the logical ZIO parents. Also, we have to add the
ZIO_FLAG_DONT_RETRY to the logical parent in zio_notify_parent(). The
reason for this is we don't want the logical ZIO parent to try and
reissue the write through zio_vdev_io_assess().

In order to make sure we returning EINVAL, I updated
manipulate_user_buffer.c to check for EINVAL if the -v flag is passed.
This can then be used to ensure we are getting EINVAL as expected.

Finally, as part of this commit Brian Behlendorf found we git the
ASSERT3U(*errrop, ==, EINVAL) with the dio_max_recordsize test case. The
error was 0, but the check was expecting EINVAL. This should not have
occured and it might be possible this happen because the child did set
the zp_direct_write_verify_error to B_FALSE in the ZIO created in
zio_vdev_child_io(). I went ahead and updated this so we make sure to
set the value in the policy for the ZIO child.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>