Fixing issues with vs_dio_verify_error
Brian Behlendorf pointed out in the PR that vs_dio_verify_error needed
to be moved to the end of the vdev_stat_t structure. Also
VDEV_STAT_VALID guard neded to be used in print_status_config() to
handle the case where new user utilities are used with old kmods.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>