Updating checks in Linux ZPL layers
Previously we were checking if !iov_iter_count() was true to return if a
O_DIRECT read/write was successful. Brian Behlendorf found this was not
entirely corret and restrictive. I updated these checks as well as the
AIO hooks with updates from Brian.

I also updated the ASSERT checks for file position as well to be less
restrictive. This allows for better sanity checks for writes.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>