Updating how to propagate checksum verify error
Commit 4938d01 updated the zio_flag to be an unit64_t. This update
allowed for us to propagate a direct I/O write checksum verify failure
up through the ZIO flags. This is a nicer implementation of this as
stashing that value in the zio_prop_t didn't really make sense as this
was used to set the write policy of ZIO LOGICAL parent and really
should not be used by ZIO VDEV child type.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>