Updating dio_write_verify test for Linux
Previously manipulate_user_buffer used either a runtime or the number of
blocks to write before stopping. However, we saw in the CI runs we would
occasionally fail this tests from time to time. Brian Behlendorf and I
looked at this and thought out why this might happen. When doing a
runtime, we are constantly just rewriting the same 8 blocks. There is a
possibility in this case we are manipulating the buffer before a
checksum is calculated.

In order to remedy this, I removed the runtime from
manipulate_user_buffer completely. Now only a certain number of blocks
will be written. This allows for a large number of blocks to be written
instead of just the last 8. Testing has shown we now reliably fail and
pass as expected.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>