Cleaning up dmu_write_direct_done()
Instead of redirtying the dirty record in the event their is an IO error
with a direct IO write, we should just do like we do on success. This
means just waiting for the other dirty records to sync out so we are
never in an inconsistent state for the TXG's and the dirty records. This
change also allows for further cleanup in the function as well.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>