Updating alignment check for iov_iter
Kernel 6.0 introduced a new iov_iter type (ITER_UBUF):
https://lore.kernel.org/all/Yp7QP9R6eDmx82GW@zeniv-ca.linux.org.uk/

Because of this our previous checks in zfs_uio_page_aligned() were
failing with a block size 512b because we were not grabbing the right
variables from the union inside of the struct iov_iter. Instead of
checking for the ITER_BUF type through iter_is_ubuf(), we can simply
just call iov_iter_alignment() and then use IS_P2ALIGNED() to check if
the struct iov_iter is page aligned. This allows us to further use the
iov_iter interfaces. This is preferable to avoid falling into the same
issue that was discovered here.

Also, there is one consequenece of using iov_iter_alignment. We do not
get the quick exit in the event any of the previous struct iovec's are
not page aligned. The kernel does have iov_iter_is_aligned() for a
quick exit, but it is a GPL exported symbol.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>