Updating and cleaning up ZTS tests
dio_aysnc_fio_engines:
Updated to make sure we test with --ionegine=io_uring if it is
available.

dio_write_verify:
Needed to update that we set zfs_vdev_direct_write_verify_pct back to
the proper default of 2. Also, I added log_note to the output for what
we are checking for so it is easier to parse failures that might have
occurred in the logs.

dio_write_stable_pages:
Added log_note to the output for what we are checking for so it is
easier to parse failures that might have occurred in the logs.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>