Removing unnecessary check in dbuf_sync_leaf()
In commit 31983d2 the code for undirtying the dbuf in
dmu_write_direct_done() was updated to just wait if there was an ARC
buffer associated with the Direct I/O write. With this update to the
wait, there is no longer a reason for us to check if we should wait in
dbuf_sync_leaf() if there is an immediate write in progress. This
intitally was needed because there was a race between dmu_buf_undirty()
and dbuf_sync_leaf(). However, we are no longer trying to undirty any of
the previous dirty records and instead are just waiting for any previous
dirty records to sync out through the TXG.

The code in dmu_write_direct_done() was further updated in e115bd4 to
also wait in the event there is an ARC buf and there is io_error. So
there should never be a reason not to wait on an immediate write as the
Direct I/O code was updated to remove any possibility of a race
condition.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>