Removing super_block flags
When removing all the code for stable pages for direct IO writes for
Linux I forgot to remove the flags for the struct super_block. This
should have be removed previously. Now all code related to trying to get
stable pages on the Linux side should be completely removed from the
code.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>