Update ZTS dio_mmap test case
89cd219 addressed an issue related to mmap'ing a file. That commit also
introduced a varient of the dio_mmap test case. In order to make sure we
are testing for the same things, I updated the dio_mmap test case to
replicate the dio_mixed test case that as added as is it a more thorough
test case than the original dio_mmap test case.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>