Fix manipulate_user_buffer code
Without using the wrote variable outside of an assert the is a warnings
(Linux/FreeBSD 14) and error (FreeBSD 13) when trying to compile without
debuging enabled through configure. This was an oversight as the wrote
value should always be checked anyways to make sure that the value
returned from pwrite is always equal to the requested blocksize.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>