Fixing Direct I/O overwrite issue
Constantly overwriting the same blocks using Direct I/O can cause a
zpool to be suspended. This happens because all Direct I/O writes occur
in open-context and DVA allocations take place when the write is
happening. Normally this is not an issue with buffered writes, because
in zfs_write() dmu_tx_assign() will return ENOSPC in the event there is
not enough space in the zpool to satisfy the write. However, the space
allocation calculation is partly based on deferred frees that take place
in the syncing context. With Direct I/O constantly overwriting and
allocating DVA's without any frees, eventually all space in the zpool is
exhausted. This leads to gang blocks trying to be used for Direct I/O
writes, but when that also fails with ENOSPC, the zpool is eventually
suspended.

In order to correct this, when a Direct I/O write is issued we can
simply check if there is an existing Direct I/O write to the same block
in the same TXG group in open-context. If there is, then the dirty
record can be undirtied and the previous DVA allocations can be added to
the defer free list. This allows for space allocation to be done
correctly for dmu_tx_asign() to operate the same for both buffered and
Direct I/O writes in zfs_write(). This also allows Direct I/O overwrites
to not exhaust all the metaslabs in the zpool.

A ZTS test was added to stress this test case with constant Direct I/O
overwrites to random offsets to the same file using FIO.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>