Adding zpool scrub to dio_verify_write test
I am temporarily adding a zpool scrub to the dio_write_verify ZTS test
case. George Amanakis is looking to the proper solution to not have to
do this. However, his commit:
431083f
Fixes in persistent error log
cause the dio_write_verify test to fail for the last test case of
compression=off and zfs_vdev_direct_write_verify_pct=100. This is due to
the fact that previous errors are still needing to be synced. Below is
his explaination:
Ok so I made that change deliberately, otherwise if we delete a file
before syncing the errlist to the disk then we cannot check if
underlying snapshots still reference that file. You can fix the test by
adding a zpool scrub -w $TESTPOOL right after the "compression=off"
finishes. That clears the error and passes the test. I will take a look
to see if we can avoid filing errors in case 1) the file has been
deleted and 2) there are no underlying snapshots and 3) the errlist has
not yet been synced to disk.

Once he has a solution, we can remove the zpool scrub from the test
case.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>