Never fault in Direct I/O write pages
In zfs_write(), zfs_uio_prefaultpages() is used to fault in any pages
that are not resident. This is an optimization, but it makes no sense to
do this for O_DIRECT pages. The reason is that pages from an O_DIRECT
write will call __get_user_pages() in the Linux kernel either through:
iov_iter_get_pages()
iov_iter_get_pages2()
get_user_pages_unlocked()
which pins the user pages into memory. Pinned pages can not get a page
fault once they are pinned in.

The pages for a O_DIRECT write are pinned through the above Linux calls
through:
zfs_setup_direct() -> zfs_uio_get_dio_page_alloc()
which happen the ZPL layer before zfs_write() is called. So this
optimization for O_DIRECT writes does not make sense to do. The faulting
in of O_DIRECT pages is now short circuited in zfs_uio_prefaultpage() by
doing nothing if the UIO has been flagged with UIO_DIRECT.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>