Properly handling larger blocks from zfs_write()
Recently b0cbc1a was merged into master. This commit allows zfs_write()
to issue more than one block at a time. Previously, this was limited to
only issueing a single block.

However, with these new changes the logic in dmu_write_uio_dnode() had
to be adjusted to propery handle this. Since more than one block can now
come, the offset might not be block aligned but the overall write is
still >= dn->dn_datablksz. In this case, we still should handle as much
of it as we can using O_DIRECT as that is what the user asked for. So we
simply will issue any misaligned (not block-size aligned) portions of
the write through the ARC and then issue the remaining that is
blocksized aligned out through O_DIRECT.

This has been verified to work as dio_aligned_block is now passing where
previously the "#large unaligned writes which span multiple blocks"
section was failing as every block was being written through the ARC.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>