Updating the code again to fix race condtion
Received further feedback from Alexander Motin on this:
Could you please explain why do you still copy the buffer before
destroying? I've written above how do I see it: just take the last dirty
record after undirtying the dbuf for this txg, and compare its buffer
with dbuf's buffer. If they are different -- free the dbuf's one, and
then in either case wipe the pointers.

Along with that explaination he provided code that made perfect sense.
This the correct way to do this. We do not need to make a copy at all.
We just need to make sure we don't destroy the ARC buffer if the
previous dirty record references it.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>