Adding tracing that solved issue with VM installs
There was an error that was occuring when calling iov_iter_advance() in
zfs_uio_get_dio_pages_iov_iter(). This was observed by doing the
follwing:
1. Base OS -> Fedora 37
2. VM OS -> Fedora 38
3. Creating dataset with direct=always
4. Created qcow2 image in that dataset
5. Used Virtual Machine Manager to install the VM OS with the disk cache
   mode set to directsynce

With all the above, the VM install would stall and an error was hit when
calling iov_iter_advance(). After looking at the call to
iov_iter_get_pages2(), on success that kernel interface calls
iov_iter_advance(). Calling it again without the configure gaurd in
place led to the issue. This was a massive oversight on my part. With
the guard, the VM installs fine. The kernel that was used on the Fedora
37 base image was 6.5.12-100.fc37.x86_64. iov_iter_get_pages2() was
introduced in kernel 6.0.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>