Fixing issue with alignment check
1. When testing out installing a VM with virtual manager on Linux and a
   dataset with direct=always, there an ASSERT failure in
   abd_alloc_from_pages(). Originally zfs_setup_direct() did an
   alignment check of the UIO using SPA_MINBLOCKSIZE with
   zfs_uio_aligned(). The idea behind this was maybe the page alignment
   restriction could be changed to use ashift as the alignment check in
   the future. Howver, this diea never came to be. The alignment
   restrictions for Direct I/O are based on PAGE_SIZE. Updating the
   check zfs_setup_direct() for the UIO to use PAGE_SIZE fixed the
   issue.
2. Updated other alignment check in dmu_read_impl() to also use
   PAGE_SIZE.
3. As a consequence of updating the UIO alignment checks the ZTS test
   case dio_unaligned_filesize began to fail. This is because there was
   no way to detect reading past the end of the file before issue EINVAL
   in the ZPL and VOPs layers in FreeBSD. This was resolved by moving
   zfs_setup_direct() into zfs_write() and zfs_read(). This allows for
   other error checking to take place before checking any Direct I/O
   limitations. Updating the call site of zfs_setup_direct() did require
   a bit of changes to the logic in that function but it mostly
   remained the same.

Originally zfs_setup_direct() did an alignment check of the UIO using
SPA_MINBLOCKSIZE with zfs_uio_aligned(). The idea behind this was maybe
the page alignment restriction could be loosed in the future to be
ashift alignment. However, this idea never came to be. Because of this,
a ASSERT failure was beeing hit in abd_alloc_from_pages() when testing
using virtual manager and installed a VM on a ZFS dataset with
direct=always:

[ 3582.257770] VERIFY0(ABD_SCATTER(abd).abd_offset) failed (0 == 512)
[ 3582.257773] PANIC at abd_os.c:911:abd_alloc_from_pages()
[ 3582.257774] Showing stack for process 60378
[ 3582.257775] CPU: 0 PID: 60378 Comm: worker Tainted: P
OE      6.5.12-100.fc37.x86_64 #1
[ 3582.257777] Hardware name: Dell Inc. Precision 5570/0T4V6W, BIOS
1.25.0 07/11/2024
[ 3582.257778] Call Trace:
[ 3582.257779]  <TASK>
[ 3582.257782]  dump_stack_lvl+0x47/0x60
[ 3582.257788]  spl_panic+0x100/0x120 [spl]
[ 3582.257802]  abd_alloc_from_pages+0x17a/0x220 [zfs]
[ 3582.258008]  dmu_read_uio_direct+0x57/0x100 [zfs]
[ 3582.258156]  zfs_read+0x238/0x3f0 [zfs]
[ 3582.258334]  zpl_iter_read_direct+0xfa/0x1a0 [zfs]
[ 3582.258498]  zpl_iter_read+0x9d/0xb0 [zfs]
[ 3582.258656]  vfs_read+0x1fe/0x350
[ 3582.258659]  __x64_sys_pread64+0x98/0xd0
[ 3582.258661]  do_syscall_64+0x5c/0x90
[ 3582.258664]  ? ksys_write+0xd8/0xf0
[ 3582.258665]  ? syscall_exit_to_user_mode+0x2b/0x40
[ 3582.258667]  ? do_syscall_64+0x6b/0x90
[ 3582.258669]  ? __x64_sys_pread64+0xa8/0xd0
[ 3582.258670]  ? syscall_exit_to_user_mode+0x2b/0x40
[ 3582.258671]  ? do_syscall_64+0x6b/0x90
[ 3582.258672]  ? syscall_exit_to_user_mode+0x2b/0x40
[ 3582.258674]  ? do_syscall_64+0x6b/0x90
[ 3582.258675]  entry_SYSCALL_64_after_hwframe+0x6e/0xd8

This also could be easily replicated by just doing FIO and passing in a
offset that was +512. This was an easy fix by just updated the uio
alignment check to use PAGE_SIZE instead. I also updated the check in
dmu_read_impl() as well. However, there was one consequence of this
change. After doing this, the test dio_unaligned_filesize began to fail.
That was because zfs_setup_direct() was being called in the zpl layers
and would return EINVAL with the new alignment check in place. This
should not occur though. To fix this issue, zfs_setup_direct() is now
called only in zfs_read() and in zfs_write(). While this maybe is not
optimal as EAGAIN may happen and the entire stack upwinds to use the
ARC, it allows for all the normal checks that take place in zfs_read()
and zfs_write() before doing Direct I/O. This is the logical way to do
things and matches other FS's as well.

So one of the good things about moving zfs_setup_direct() inside of
zfs_read() and zfs_write() is the rangelocks could "in theory" be held
before doing checks if the file region is mmap'ed. However, there is an
issue that occurred when doing this. We don't want to prefault any
Direct I/O pages as those will be pinned in the kernel and can not be
faulted out at any point. When moving zfs_setup_direct() and
zfs_uio_prefault() after grabbing the rangelock in zfs_write() certain
ZTS mmap tests would wind up not going anywhere and be KILLED after 10
mins.

Even in the cases of the mmap test cases not being KILLED, adding the
ASSERT0(ioflag & O_DIRECT) before calling update_pages() in zfs_write
would fail with the dio_mmap case. This is really strange, as we should
not get into this situation while holding the rangelock. When dolding
that and doing the same zn_has_cached_data() over the same region, it
should always be either true or false in both zfs_setup_direct() and
before calling update_pages() in zfs_write(). Unfortunatly, this is not
the case and that means there is a place where mapping in pages in Linux
is not protected by a rangelock... No matter what though, it does not
matter if we can not call zfs_uio_prefaultpages() while holding the
rangelock. zfs_uio_prefaultpages() should be called after
zfs_setup_direct() as the UIO_DIRECT flag is needed to determine if the
pages are pinned in memory to avoid doing the costly fautling in of
pages that would not be necessary for Direct I/O writes.

The logic in zfs_setup_direct() did need to change a bit with moving
it's call site, but the logic remained mostly the same.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>