Added tracing and found deadlock
I added zfs_dbgmsg() statements in zfs_write() and zfs_getpage() to see
if there was a deadlock while holding the page lock. Sure enough there
is. As part of the O_DIRECT code, grabbing the rangelock as reader in
zfs_getpage() was added. This is required and described in detail the
comments in that code. In theory, we should also be able to call
zfs_uio_prefaultpages() while holding the rangelock. In fact we already
do this at line 821. However, this assumption is now wrong as we are
attempting to hold the rangelock in zfs_getpage() over a region that is
mmaped. This means when zfs_uio_prefaultpages() goes to pefault that
region in, it will call zfs_getpage(), which leads to a deadlock. I
tried moving both calls of zfs_uio_prefaultpages() under the rangelock
in zfs_write() and hit the deadlock.

This was observed while running the ZTS mmap test cases and was hit in
mmap_libaio_001_pos. Below are the zfs_dbgmsg() statements showing the
deadlock occur for znode z_id 2:
1724949914   ffff8c3c8fcb8000 zfs_vnops.c:692:zfs_write(): Going to prefault pages for znode: 2 while holding rangelock as RL_WRITER over woff = 0, n = 512, pfbytes = 512
1724949914   ffff8c3c8fcb8000 zfs_vnops_os.c:4087:zfs_getpage(): Using zfs_rangelock_tryenter() on znode: 2, io_off = 0, io_len = 512, as a RL_READER
1724949914   ffff8c3c8fcb8000 zfs_vnops_os.c:4105:zfs_getpage(): Trying to acquire rangelock failed. Dropped page lock and now using zfs_rangelock_enter() on znode: 2, io_off = 0, io_len = 512, as a RL_READER

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>