Trying to use rangelocks to detect mmap
I updated the rangelock in zfs_getpages() to a RL_WRITER lock. The
reasoning behind this was that then in zfs_read(), the check for
zn_has_cached_data() would be checked while holding the rangelock as
reader. So, the region should only be mapped or not. This would fix a
race condition. This also is important because we call
zfs_setup_direct() in zfs_read() holding the rangelock already. Yet, for
some strange reason, we are hitting the ASSERT0(uio->uio_extflg &
UIO_DIRECT)... There still exists a race condition even while holding
the rangelock as a writer in zfs_getpage().

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>